ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pytesseract==0.3.10
Pillow==10.1.0
pdfplumber==0.10.3
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99


